# cooperation/src/main/resources/static/doc/openapi_v2.yaml
openapi: 3.0.1
info:
  title: Etablerad samverkan i vården
  description: "![Informationsmodell - Cooperation](images/cooperation.png)\r\n\r\nInformation om tjänster, producenter och konsumenter baserad på tjänsteadresseringskatalogen i både Ineras nationella tjänsteplattform och regionala tjänsteplattformar. Eventuella framtida konfigurationsändringar syns inte i api:et, enbart hur det ser ut i dagsläget. Konfiguration i en tjänsteplattform kan ibland ske på en viss nivå i HSA-trädet, och avser alla noder därunder. Detta speglas rakt av i api:et och HSA-trädet evalueras inte här. De tekniska id-begreppen på alla objekt genereras om en gång per dygn och garanteras således inte ha samma värde två dagar i rad. Applikationer bör skrivas så att de är oberoende av de tekniska nycklarna. Varje resurs har sin egen nummerserie för de tekniska id-begreppen. En api-nyckel kan behövas för att nå api:et."
  version: 2.0.0
servers:
  - url: /coop/api/v2
paths:
...
  /serviceConsumers:
    get:
      summary: Service Consumer - tjänstekonsument
      description: |
        Alla tjänstekonsumenter. En konsument identifieras versamhetsmässigt av hsa-id. En förekomster av konsument i detta api representerar en konsument i en instans instanser av tjänsteplattformen. slås ihop till ett objekt. Samma hsa-id kan alltså förekomma i mer är ett objekt. Om connectionPoint inte anges som query-parameter så inkluderas connectionPoint till varje objekt i svaret  
      parameters:
        - name: connectionPointId
          in: query
          description: Filtrerar bara tjänstekonsumenter som hör till den aktulla instansen av tjänsteplattformen. Ange den tekniska nyckeln för connectionPoint tex 376.
          required: false
          schema:
            type: integer
            format: int64
        - name: logicalAddressId
          in: query
          description: Filtrerar bara tjänstekonsumenter som har någon samverkan med en specifik logisk adress. Ange den tekniska nyckeln för LogicalAddress tex 42.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceContractId
          in: query
          description: Filtrerar bara tjänstekonsumenter som har någon samverkan med ett specifik tjänstekontrakt. Ange den tekniska nyckeln för ServiceContract tex 123.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceProducerId
          in: query
          description: Filtrerar bara tjänstekonsumenter som har någon samverkan med en specifik tjänsteproducent. Sökningen utförs genom att titta på tjänstproducenter som publicerar en tjänst på en viss logisk adress samtidigt som någon tjänstekonsument är konfigurerad att anropa samma logiska adress.  Ange den tekniska nyckeln för ServiceProducer tex 538.
          required: false
          schema:
            type: integer
            format: int64
      tags:
        - ServiceConsumer
      responses:
        '200':
          description: En lista av tjänstekonsumenter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceConsumer'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceConsumer'
                xml:
                  wrapped: true
                  name: serviceConsumers
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
  /serviceConsumers/{id}:
    get:
      summary: Service Consumer - tjänstekonsument
      description: |
        En specifik tjänstekonsument.
      parameters:
        - $ref: '#/components/parameters/TechnicalKeyParam'
      tags:
        - ServiceConsumer
      responses:
        '200':
          description: En tjänstekonsument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceConsumer'
            application/xml:
              schema:
                $ref: '#/components/schemas/ServiceConsumer'
        '404':
          description: ServiceConsumer not found
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
...
  /installedContracts:
    get:
      summary: InstalledContract - installerade tjänstekontrakt
      description: |
        Alla tjänstekontrakt som är installerade i en viss miljö.  Kontraktet kan vara installerat även om det inte finns några kopplingar mot det ännu  
      parameters:
        - name: connectionPointId
          in: query
          description: Filtrerar installerade tjänstekontrakt som finns i den aktulla instansen av tjänsteplattformen. Ange den tekniska nyckeln för connectionPoint tex 376.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceContractId
          in: query
          description: Filtrerar bara installerade tjänstekontrakt av en viss typ. Ange den tekniska nyckeln för ServiceContract tex 123.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceDomainId
          in: query
          description: Filtrerar bara installerade tjänstekontrakt avseende kontrakt som ingår i en specifik tjänstedomän.  Ange den tekniska nyckeln för ServiceDomain tex 538.
          required: false
          schema:
            type: integer
            format: int64
      tags:
        - InstalledContract
      responses:
        '200':
          description: En lista av installerade tjänstekontrakt
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstalledContract'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstalledContract'
                xml:
                  wrapped: true
                  name: installedContracts
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
  /installedContracts/{id}:
    get:
      summary: InstalledContract - installerade tjänstekontrakt
      description: |
        En specifik installation av ett tjänstekontrakt i en viss miljö. Returobjekten inkluderar alltid ConnectionPoint och ServiceContract, annars hade det bara blivit en lista av tekniska nycklar.
      parameters:
        - $ref: '#/components/parameters/TechnicalKeyParam'
      tags:
        - InstalledContract
      responses:
        '200':
          description: Ett installerat tjänstekontrakt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstalledContract'
            application/xml:
              schema:
                $ref: '#/components/schemas/InstalledContract'
        '404':
          description: InstalledContract not found
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:     # <-- arbitrary name for the security scheme
      type: http
      scheme: basic
  responses:
    UnexpectedErrorResponse:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
        application/xml:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    TechnicalKeyParam:
      name: id
      in: path
      description: Teknisk nyckel
      required: true
      schema:
        type: number
  schemas:
    ConnectionPoint:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        platform:
          type: string
          description: Plattform tex NTJP.
        environment:
          type: string
          description: Miljö tex PRODUCTION.
        snapshotTime:
          type: string
          description: Den tidpunkt som informationen lästes av från tjänsteplattformen.
      xml:
        name: connectionPoint
    ServiceConsumer:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        hsaId:
          type: string
          description: hsa-id.
        description:
          type: string
          description: Beskrivning.
      xml:
        name: serviceConsumer
...
    ServiceContract:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        name:
          type: string
          description: Namn.
        namespace:
          type: string
          description: Namnrymd.
        major:
          type: number
          description: Major-version exvis 1 eller 2.
        minor:
          type: number
          description: Minor-version exvis 0 eller 1.
      xml:
        name: serviceContract
...
    InstalledContract:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        connectionPoint:
          $ref: '#/components/schemas/ConnectionPoint'
        serviceContract:
          $ref: '#/components/schemas/ServiceContract'
      xml:
        name: installedContract
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        fields:
          type: string
      xml:
        name: error